a0.2.10.44 - Hotfix patch №2

+++ CloneDroneOverhaulBeta/Internal/AssetsManagement/Network/OverhaulNetworkController.cs
@@ -13,7 +13,7 @@ namespace CDOverhaul.NetworkAssets
         public static string DownloadFolder => OverhaulMod.Core.ModDirectory + "Assets/Download/";
         public static string PermanentDownloadFolder => OverhaulMod.Core.ModDirectory + "Assets/Download/Permanent/";
 
-        public static float MultiplayerLocalPing => MultiplayerPlayerInfoManager.Instance == null ? 0f : (float)MultiplayerPlayerInfoManager.Instance.GetLocalPing() / 1000f;
+        public static float MultiplayerLocalPing => MultiplayerPlayerInfoManager.Instance == null ? 0f : MultiplayerPlayerInfoManager.Instance.GetLocalPing() / 1000f;
 
         public static OverhaulNetworkDownloadHandler DownloadAndSaveFile(string address, string directoryPath, string fileName, Action onDoneDownloadingAction)
         {
